% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyRisk.R
\name{cal_rp}
\alias{cal_rp}
\title{A function evaluting a conditional mean (CM) of being affected of the disease using family history}
\usage{
cal_rp(fam, h, prev, target)
}
\arguments{
\item{fam}{Input file in FAM format of PLINK. Target individuals and their relatives should be all included in the FAM file.}

\item{h}{Heritability of the disease}

\item{prev}{Prevalence of the disease (should be specified according to the population)}

\item{target}{Target individuals whose conditional mean being evaluated. This should be a index vector in the samples in the FAM file}
}
\value{
\item{res_rp_I}{a matrix including IID and CM}
}
\description{
A function evaluting a conditional mean (CM) of being affected of the disease using family history
}
\details{
Please see the reference paper.
}
\examples{
(you example code, each line start with #')
}
\references{
Jungsoo Gim, 2017, Genetics, "Incorporating family disease history in risk prediction models with large-scale genetic data substantially reduces unexplained variation."
}
\author{
Jungsoo Gim
}
